requirements.txt
fastapi
uvicorn
structlog
pydantic-settings
pytest
